+++ tasks/chyokotov_a_seidel_method/tests/functional/main.cpp
@@ -34,7 +34,7 @@ class ChyokotovASeidelMethodFuncTest : public ppc::util::BaseRunFuncTests<InType
   }
 
   bool CheckTestOutputData(OutType &output_data) final {
-    for (int i = 0; i < output_data.size(); ++i) {
+    for (size_t i = 0; i < output_data.size(); ++i) {
       if (std::abs(output_data[i] - expected_output_[i]) > 0.001) {
         return false;
       }
